+++ CHANGES.md
@@ -14,3 +14,6 @@
 * Fixed error "Column reference `res_name` is ambiguous" in report
   `erm_agreement_package_content_item_list`.
 
+* Updated report query `erm_agreement_cancellation_dates` with
+  improvements and to support schema changes in the source data.
+